added lib/events.js and lib/selection.js
